moved the empty_search_term and contain_symbol methods into the Confirmation module and am including the module in the Edamama wrapper class. But now I'm not sure Iwant to do this, because I don't actually want to raise errors in the controller, which is what these methods do...
