--- lib/confirmation_module.rb
@@ -1,1 +1,0 @@
-
